Initial changes to new hierarchy.
This is all in an attempt to remove any dependence of actions on dataset standard. There were several places where special functionality was added by checking if the dataset is a prtDataSetStandard. These cases have been removed by adding special methods in prtDataSetBase that can be overloaded appropriately. Partial testing is complete. Actions have temporarily lost the ability to modify feature names. This may also have broken backwards compatibility for loading and saving as a few properties now no longer exist. Most notably verboseFeatureNames is no longer a property of actions. It will eventually be a property of prtDataSetStandard.